GenAI-Project/requirements.txt
python-dotenv
openai>=1.0.0
